dotty.ttyctl: Import basic module to control a TTY
This imports the dotty.ttyctl which acts as low-level TTY controller.
For the moment it is only able to set cbreak mode and restore the
terminal back from it.

As a convenience, the ttyctl:with_cbreak() method will run a function in
cbreak mode, catch errors (it any) and then return the console to normal
mode, printing errors that may have been produced while in cbreak mode.
This can be used as follows to read one byte of input from the keyboard
without waiting the "enter" key to be pressed:

  function getch()
    local c
    ttyctl(io.stdout):with_cbreak(function ()
      c = io.read(1)
    end)
    return c
  end

For a more comprehensive example, check examples/ttyctl-getch.lua