First pass at the terminal escape sequence parser
The dotty.descape module implements a parser for terminal escape sequences.
For the moment, only some CSI sequences are parsed, nevertheless the module
is already a working solid foundation.

An small example script (examples/descape-delegate.lua) which exercises the
module in combination with dotty.unidecode and dotty.ttyctl is provided as
well.