descape: Unify CSI and non-CSI function key handling
This reuses the csi_add_modifier_flags() function in the handlers for VT52
and VT100/ANSI function key escape sequences. This also makes the function
more composable, by returning all the parameters which are to be passed to
the delegate handlers. As a side effect, the manual unpack() is not needed
and the values returned by mangling functions in tables do not need to modify
the received parameters in-place.

Finally, this makes a second argument to be passed to the delegate handlers,
which is always a number: for movement keys (up, down, left, right) it is the
number of positions moved, for other keys it is a key-specific key code.