Parse valid Alt+<key> non-CSI escape sequences
Fixes #2.

Implement recognition of Alt+<key> escape sequences. These are sent by the
terminal as "ESC letter", which unfortunately overlaps for the simple VT-52
escape sequences for the keypad and F1-F4 keys. This means that certain key
combinations cannot be reliably detected without disabling VT-52 mode on
terminals which support alternate keypad reporting, neither of which dotty
can do at the moment:

* Alt+Shift+A (VT-52 "up")
* Alt+Shift+B (VT-52 "down")
* Alt+Shift+C (VT-52 "right")
* Alt+Shift+D (VT-52 "left")
* Alt+Shift+F (VT-52 "end")
* Alt+Shift+H (VT-52 "home")
* Alt+Shift+O (prefix for keypad escapes in ANSI mode)
* Alt+Shift+P (VT-52 "F1")
* Alt+Shift+Q (VT-52 "F2")
* Alt+Shift+R (VT-52 "F3")
* Alt+Shift+S (VT-52 "F4")

Note that even if VT-52 mode could be disabled, it would not be possible to
handle Alt+Shift+O, as it clashes with the prefix for keypad escape sequences
when terminals use ANSI-style keypad reporting.

At any rate, probably it is a good idea for developers to avoid using any
of the above in their applications, to ensure that they will work in those
terminals which use VT-52 and/or ANSI-style keypad reporting.