Implement handling of escapes for Ctrl+Alt+<key>
This patch partially solves issue #2 for the most common use cases.

Ctrl+Alt+<key> sequences are actually sent by terminals as Alt+(Ctrl+<key>):
the terminal sends ESC first, followed by the same one byte as entering
Ctrl+<key> would produce. These are summarized in the following table:

  http://vt100.net/docs/vt102-ug/figure4-3.html

This commit implements *only* handling an alphabetic <key> (A-Z) because it
is unclear how to handle characters which change positions in international
keyboards. Most likely the generated codes depend on the character on the key
cap, and not on the physical positions of the keys, but I would like to do a
bit of investigation beforehand. In the meanwhile, this covers most of the
use cases where an application may want to do Alt+Ctrl+[A-Z] keybindings.

Also, the contained changes indirectly fix incorrect (and unneeded) handling
of a CAN/SUB byte following ESC in "descape.decode_escape()".